CRM-5 refactor: Organize Enums and Traits into dedicated directories
Description:
Refactored the project structure by moving Enums and Traits out of the Models directory into dedicated directories for better organization and maintainability.

Changes:
- Created `app/Enums` directory and moved `Can` Enum into it.
- Created `app/Traits` directory and moved `HasPermissions` trait into it.
- Updated all references to the `Can` Enum and `HasPermissions` trait throughout the codebase to reflect the new paths.

Testing:
- Verified that all existing tests pass and the application functions correctly after the refactor.
- Ensured that the new directory structure enhances clarity and organization without affecting functionality.
