+++ database/factories/UserFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\Models\Can;
+use App\Enums\Can;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\Hash;
