+++ database/seeders/UserSeeder.php
@@ -2,7 +2,7 @@
 
 namespace Database\Seeders;
 
-use App\Models\Can;
+use App\Enums\Can;
 use App\Models\User;
 use Illuminate\Database\Seeder;
